+++ lib/selectors/aggregate-accessibility.js
@@ -3,6 +3,7 @@ import {createSelector} from 'reselect'
 import {activeOpportunityDatasetGrid} from 'lib/modules/opportunity-datasets/selectors'
 import getAggregateAccessibility from 'lib/utils/aggregate-accessibility'
 
+import {active} from './regional-grid'
 import selectActiveAggregationArea from './active-aggregation-area'
 
 /**
@@ -10,7 +11,7 @@ import selectActiveAggregationArea from './active-aggregation-area'
  * analysis
  */
 export default createSelector(
-  state => state.analysis.regional.grid,
+  active,
   selectActiveAggregationArea,
   activeOpportunityDatasetGrid,
   getAggregateAccessibility
