+++ lib/selectors/comparison-aggregate-accessibility.js
@@ -3,11 +3,12 @@ import {createSelector} from 'reselect'
 import {activeOpportunityDatasetGrid} from 'lib/modules/opportunity-datasets/selectors'
 import getAggregateAccessibility from 'lib/utils/aggregate-accessibility'
 
+import {comparison} from './regional-grid'
 import selectActiveAggregationArea from './active-aggregation-area'
 
 /** Aggregate accessibility for the comparison regional analysis */
 export default createSelector(
-  state => state.analysis.regional.comparisonGrid,
+  comparison,
   // aggregation area and weights don't vary between base and comparison
   selectActiveAggregationArea,
   activeOpportunityDatasetGrid,
