+++ lib/selectors/regional-comparison-cutoff.js
@@ -13,6 +13,6 @@ export default createSelector(
     const cutoffs = get(analysis, 'cutoffsMinutes', [])
     if (cutoffs.indexOf(comparison) >= 0) return comparison
     if (cutoffs.indexOf(cutoff) >= 0) return cutoff
-    return cutoffs[0]
+    return cutoffs[Math.floor(cutoffs.length / 2)]
   }
 )
