+++ lib/selectors/regional-comparison-percentile.js
@@ -13,6 +13,6 @@ export default createSelector(
     const percentiles = get(analysis, 'travelTimePercentiles', [])
     if (percentiles.indexOf(comparison) >= 0) return comparison
     if (percentiles.indexOf(percentile) >= 0) return percentile
-    return percentiles[0]
+    return percentiles[Math.floor(percentiles.length / 2)]
   }
 )
