+++ lib/selectors/regional-display-cutoff.js
@@ -9,6 +9,6 @@ export default createSelector(
   (analysis, cutoff) => {
     const cutoffs = get(analysis, 'cutoffsMinutes', [])
     if (cutoffs.indexOf(cutoff) >= 0) return cutoff
-    return cutoffs[0]
+    return cutoffs[Math.floor(cutoffs.length / 2)]
   }
 )
