+++ lib/selectors/regional-display-grid.js
@@ -1,42 +1,12 @@
-import get from 'lodash/get'
 import {createSelector} from 'reselect'
 
-import selectActiveAnalysis from './active-regional-analysis'
 import selectComparisonAnalysis from './comparison-regional-analysis'
-import selectComparisonCutoff from './regional-comparison-cutoff'
-import selectComparisonPercentile from './regional-comparison-percentile'
-import selectDisplayCutoff from './regional-display-cutoff'
-import selectDisplayPercentile from './regional-display-percentile'
-
-const findGrid = (grids, analysis, cutoff, percentile) =>
-  analysis &&
-  grids.find(
-    g =>
-      g.analysisId === analysis._id &&
-      g.cutoff === cutoff &&
-      g.percentile === percentile
-  )
-
-const selectGrid = createSelector(
-  state => get(state, 'regionalAnalyses.grids'),
-  selectActiveAnalysis,
-  selectDisplayCutoff,
-  selectDisplayPercentile,
-  findGrid
-)
-
-const selectComparisonGrid = createSelector(
-  state => get(state, 'regionalAnalyses.grids'),
-  selectComparisonAnalysis,
-  selectComparisonCutoff,
-  selectComparisonPercentile,
-  findGrid
-)
+import * as selectGrid from './regional-grid'
 
 export default createSelector(
-  selectGrid,
+  selectGrid.active,
   selectComparisonAnalysis,
-  selectComparisonGrid,
+  selectGrid.comparison,
   (activeGrid, comparisonAnalysis, comparisonGrid) => {
     if (comparisonAnalysis) {
       if (activeGrid && comparisonGrid)
