+++ lib/selectors/regional-display-percentile.js
@@ -9,6 +9,6 @@ export default createSelector(
   (analysis, percentile) => {
     const percentiles = get(analysis, 'travelTimePercentiles', [])
     if (percentiles.indexOf(percentile) >= 0) return percentile
-    return percentiles[0]
+    return percentiles[Math.floor(percentiles.length / 2)]
   }
 )
